Split into components. Your balance restyled
