+++ webapp/src/components/Icon.css
@@ -23,11 +23,14 @@
   background-image: url("../images/laquo.svg");
 }
 
-.Icon-Shift {
+.Icon-shift {
   width: 11px;
   height: 13px;
-  margin-right: 6px;
-  position: relative;
-  top: 1px;
   background-image: url("../images/shift.svg");
+}
+
+.Icon-address {
+  width: 20px;
+  height: 20px;
+  background-image: url("../images/address.svg");
 }
